Move the patch to an autouse fixture
